refactor repo structure
